# Improve performance of answerGetBlockBodiesQuery