--- a/eth/protocols/eth/handlers.go
+++ b/eth/protocols/eth/handlers.go
@@ -153,14 +153,15 @@ func answerGetBlockBodiesQuery(backend Backend, query GetBlockBodiesPacket, peer
 			lookups >= 2*maxBodiesServe {
 			break
 		}
-		// Retrieve the requested block body, stopping if enough was found
-		if body := backend.Chain().GetBody(hash); body != nil {
-			bh := &blockBodyWithBlockHash{BlockHash: hash, BlockBody: body}
-			bhRLPbytes, err := rlp.EncodeToBytes(bh)
+		// Retrieve the requested block body RLP, stopping if enough was found
+		if body := backend.Chain().GetBodyRLP(hash); len(body) != 0 {
+			hashRLP, err := rlp.EncodeToBytes(hash)
+
 			if err != nil {
 				return nil, err
 			}
-			bhRLP := rlp.RawValue(bhRLPbytes)
+			bhRLP := rlp.Combine(hashRLP, body)
+
 			bodiesAndBlockHashes = append(bodiesAndBlockHashes, bhRLP)
 			bytes += len(bhRLP)
 		}