### Description

Avoid decoding and re-encoding of the block body by using available RLP-encoded block body from the database, combine it with the RLP-encoded block hash to avoid unnecessary processing and extra memory allocations.

![Screenshot 2023-06-26 at 17 36 19](https://github.com/celo-org/celo-blockchain/assets/20768968/aab56878-88f6-433e-9ffc-a15ef6c6d7d4)

Benchmarks:
```
Original Ethereum version (highly efficient as block body is taken either from the cache or database directly).

BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-1-8         	14819121	        92.65 ns/op	      24 B/op	       1 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-10-8        	 1318404	       896.9 ns/op	     744 B/op	       5 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-100-8       	  153598	      7907 ns/op	    6120 B/op	       8 allocs/op


Current Celo-version

BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-1-8         	 1000000	      1278 ns/op	     448 B/op	      12 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-10-8        	   81495	     12543 ns/op	    4984 B/op	     115 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-100-8       	   10000	    115520 ns/op	   48523 B/op	    1108 allocs/op


Suggested version

BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-1-8         	 3173032	       370.5 ns/op	     256 B/op	       6 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-10-8        	  297216	      3805 ns/op	    3064 B/op	      55 allocs/op
BenchmarkAnswerGetBlockBodiesQuery/GetBlockBodies-100-8       	   30360	     37067 ns/op	   29322 B/op	     508 allocs/op
```

3x more CPU-efficient and 1.5x more memory-efficient.
A larger change would be storing RLP-encoded blockBodyWithBlockHash structures instead of just block bodies but it would significantly increase the diff.

### Other changes

No.

### Tested

Existing tests pass, added unit tests and the benchmark.

### Backwards compatibility

Yes.